Extract peer-link handshake driver into handshake.py
Fifth PR in the peer_link split arc (follows #769 + #770 +
#773 + #781). Moves the Noise XX handshake driver +
intent-dispatch unit byte-for-byte into peer_link/handshake.py:

* Private types: _HandshakeStep, _DispatchInput
* Driver: _drive_peer_link_session
* Dispatcher: _dispatch_intent

The handshake driver's only callback into the parent package
is _run_peer_link_session (still in __init__.py until PR 6
extracts the session machinery); imported lazily inside the
peer-link-intent branch to avoid the circular-load that a
top-level import would trigger. Same pattern channel.py used
for its parse_app_frame lazy import.

__init__.py re-exports the four handshake symbols (PEP 484
redundant-alias form) so existing test imports stay valid.
wire_io.py's TYPE_CHECKING reference to _HandshakeStep follows
the symbol from the package to the handshake submodule.

Drops the now-unused top-level imports from __init__.py that
were only consumed by the moved code (StrEnum, dataclass+field
stayed for PeerLinkSession, DASHBOARD_ID_*, HandshakeNotCompleteError,
pin_sha256_for_pubkey, IntentResponse, PeerLinkIntent).

Test redirects: none. _drive_peer_link_session is still
called from make_peer_link_handler (which stays in
__init__.py), so the existing
peer_link._drive_peer_link_session monkeypatch target is
unchanged — patching the package-level re-export reaches
make_peer_link_handler's own scope inside __init__.py.

__init__.py: 729 -> 519 lines (-210). handshake.py: 240 lines.

All 3391 non-e2e tests pass; ruff + mypy clean.
